+++ master-nafsdm/manager/daemonlog.py
@@ -8,7 +8,7 @@ from time import strftime
 
 def log(log):
     # write it to a file with current time / date
-    write = strftime("%a, %d %b %Y %H:%M:%S") + " - " + log + "\n"
+    write = "MANAGER " + strftime("%a, %d %b %Y %H:%M:%S") + " - " + log + "\n"
 
     f = open("/home/master-nafsdm/log.log", "a")
     f.write(write)